Begun on the most promising solution yet (details below for archiving purposes):
Basically all touch events (MotionEvents) that are sent to the TaskListView get intercepted in TaskListView.onInterceptTouchEvent for us to be able to manipulate. Here we will use the event to make the DragShadow visible, match the Task that the event is over (if it is over one) and then tell the framework that we will handle all further events in TaskListView by returning true in TaskListView.onInterceptTouchEvent. When that happens, all further events go to TaskListView.onTouchEvent, this way we can just send them to the DragShadow, we can also send them to the touched Task if we like, or not, up to us. This means that the MotionEvent's target will always be the TaskListView, so no matter what happens to Tasks, one of them gets redrawn, detached, reattached or whatever, the MotionEvent will live on as long as the TaskListView is doing alright. The difference in this implementation is that the touched Task NEVER actually receives its MotionEvents directly, only through someone else, and so this way the system thinks the target of the MotionEvent is the TaskListView, when in actuality we're doing the correct sending ourselves so that even if the Task is redrawn or whatever, the event will still continue coming and not getting ACTION_CANCEL ed by the system.

I'll have to write about this somewhere because it has taken me a while to discover and understand this and likely it has affected others and will definitely take a while for even me to understand again, it needs to be documented and explained clearly!

Funnily enough, I had thought of a solution like this very early on, but my knowledge of the Android touch system was insufficient so what I had tried at the time did not work, but now that I know what is going on underneath, it seems that this solution might be the one that works! I'll keep working on it and see where we can go but it looks like the roadblock I've had for months is finally opening up! ^_^